backend/requirements.txt
# Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-nested-routers>=0.93,<1.0

# Database
# psycopg2-binary>=2.9,<3.0  # Uncomment for PostgreSQL

# CORS
django-cors-headers>=4.3,<5.0

# Filtering
django-filter>=23.5,<24.0

# Environment variables
python-dotenv>=1.0,<2.0

# Development
ipython>=8.0,<9.0
